hotpin-webserver/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
websockets==12.0
httpx==0.25.2
Pillow==10.1.0
vosk==0.3.45
pyttsx3==2.90

# Optional discovery dependencies
# Install with pip install .[discover] or pip install -r requirements.txt
psutil>=5.9.0  # For network interface listing
zeroconf>=0.132.0  # For mDNS advertisement
qrcode[pil]>=7.4.2  # For QR code generation